+++ sdk/translation/azure-ai-translation-document/src/main/java/com/azure/ai/translation/document/DocumentTranslationClient.java
@@ -90,6 +90,7 @@ public final class DocumentTranslationClient {
      *                  (Required){
      *                     targetUrl: String (Required)
      *                     category: String (Optional)
+     *                     deploymentName: String (Optional)
      *                     language: String (Required)
      *                     glossaries (Optional): [
      *                          (Optional){
@@ -105,6 +106,9 @@ public final class DocumentTranslationClient {
      *             storageType: String(Folder/File) (Optional)
      *         }
      *     ]
+     *     options (Optional): {
+     *         translateTextWithinImage: Boolean (Optional)
+     *     }
      * }
      * }
      * </pre>
@@ -255,6 +259,9 @@ public final class DocumentTranslationClient {
      *         notYetStarted: int (Required)
      *         cancelled: int (Required)
      *         totalCharacterCharged: long (Required)
+     *         totalImageScansSucceeded: Integer (Optional)
+     *         totalImageScansFailed: Integer (Optional)
+     *         totalImageCharged: Long (Optional)
      *     }
      * }
      * }
@@ -303,6 +310,11 @@ public final class DocumentTranslationClient {
      *     progress: double (Required)
      *     id: String (Required)
      *     characterCharged: Integer (Optional)
+     *     totalImageScansSucceeded: Integer (Optional)
+     *     totalImageScansFailed: Integer (Optional)
+     *     imageCharged: Integer (Optional)
+     *     imageCharacterDetected: Integer (Optional)
+     *     deploymentName: String (Optional)
      * }
      * }
      * </pre>
@@ -358,6 +370,9 @@ public final class DocumentTranslationClient {
      *         notYetStarted: int (Required)
      *         cancelled: int (Required)
      *         totalCharacterCharged: long (Required)
+     *         totalImageScansSucceeded: Integer (Optional)
+     *         totalImageScansFailed: Integer (Optional)
+     *         totalImageCharged: Long (Optional)
      *     }
      * }
      * }
@@ -416,6 +431,9 @@ public final class DocumentTranslationClient {
      *         notYetStarted: int (Required)
      *         cancelled: int (Required)
      *         totalCharacterCharged: long (Required)
+     *         totalImageScansSucceeded: Integer (Optional)
+     *         totalImageScansFailed: Integer (Optional)
+     *         totalImageCharged: Long (Optional)
      *     }
      * }
      * }
@@ -558,6 +576,11 @@ public final class DocumentTranslationClient {
      *     progress: double (Required)
      *     id: String (Required)
      *     characterCharged: Integer (Optional)
+     *     totalImageScansSucceeded: Integer (Optional)
+     *     totalImageScansFailed: Integer (Optional)
+     *     imageCharged: Integer (Optional)
+     *     imageCharacterDetected: Integer (Optional)
+     *     deploymentName: String (Optional)
      * }
      * }
      * </pre>